refactor(fs): extract syncPart2 core detectors
