+++ core/facts/detectors/fs/syncPart2.ts
@@ -1,74 +1,4 @@
-import { hasNode, isObject } from '../utils/astHelpers';
-
+export * from './syncPart2Core';
 export * from './syncPart2Permissions';
 export * from './syncPart2Io';
 export * from './syncPart2Times';
-
-export const hasFsUnlinkSyncCall = (node: unknown): boolean => {
-  return hasNode(node, (value) => {
-    if (value.type !== 'CallExpression') {
-      return false;
-    }
-
-    const callee = value.callee;
-    if (!isObject(callee) || callee.type !== 'MemberExpression' || callee.computed === true) {
-      return false;
-    }
-    const objectNode = callee.object;
-    const propertyNode = callee.property;
-    return (
-      isObject(objectNode) &&
-      objectNode.type === 'Identifier' &&
-      objectNode.name === 'fs' &&
-      isObject(propertyNode) &&
-      propertyNode.type === 'Identifier' &&
-      propertyNode.name === 'unlinkSync'
-    );
-  });
-};
-
-export const hasFsRmdirSyncCall = (node: unknown): boolean => {
-  return hasNode(node, (value) => {
-    if (value.type !== 'CallExpression') {
-      return false;
-    }
-
-    const callee = value.callee;
-    if (!isObject(callee) || callee.type !== 'MemberExpression' || callee.computed === true) {
-      return false;
-    }
-    const objectNode = callee.object;
-    const propertyNode = callee.property;
-    return (
-      isObject(objectNode) &&
-      objectNode.type === 'Identifier' &&
-      objectNode.name === 'fs' &&
-      isObject(propertyNode) &&
-      propertyNode.type === 'Identifier' &&
-      propertyNode.name === 'rmdirSync'
-    );
-  });
-};
-
-export const hasFsFstatSyncCall = (node: unknown): boolean => {
-  return hasNode(node, (value) => {
-    if (value.type !== 'CallExpression') {
-      return false;
-    }
-
-    const callee = value.callee;
-    if (!isObject(callee) || callee.type !== 'MemberExpression' || callee.computed === true) {
-      return false;
-    }
-    const objectNode = callee.object;
-    const propertyNode = callee.property;
-    return (
-      isObject(objectNode) &&
-      objectNode.type === 'Identifier' &&
-      objectNode.name === 'fs' &&
-      isObject(propertyNode) &&
-      propertyNode.type === 'Identifier' &&
-      propertyNode.name === 'fstatSync'
-    );
-  });
-};
